LRDOCS-9470 changed directory structure and renamed methods
