+++ docs/dxp/latest/en/developing-applications/core-frameworks/configurable-application/providing-dynamic-dropdown-values/resources/liferay-z4h3.zip/z4h3-impl/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Acme Z4H3 Impl
+Bundle-Name: Acme Z4H3 Implementation
 Bundle-SymbolicName: com.acme.z4h3.impl
 Bundle-Version: 1.0.0
