+++ docs/dxp/latest/en/developing-applications/core-frameworks/configurable-application/providing-dynamic-dropdown-values/resources/liferay-z4h3.zip/z4h3-impl/src/main/java/com/acme/z4h3/internal/configuration/Z4H3Configuration.java
@@ -4,24 +4,23 @@ import aQute.bnd.annotation.metatype.Meta;
 
 @Meta.OCD(
 	id = "com.acme.z4h3.internal.configuration.Z4H3Configuration",
-	localization = "content/Language", name = "z4h3-configuration"
+	localization = "content/Language", name = "z4h3-configuration-name"
 )
 public @interface Z4H3Configuration {
 
 	@Meta.AD(
-		description = "this-list-is-manually-populated",
-		name = "z4h3-asset-types",
-		optionLabels = {"Option 1", "Option 2", "Option 3"},
-		optionValues = {"option1", "option2", "option3"}, required = false
+		description = "this-list-is-manually-populated", name = "asset-types",
+		optionLabels = {"Asset1", "Asset2", "Asset3"},
+		optionValues = {"Asset1", "Asset2", "Asset3"}, required = false
 	)
-	String entryList1();
+	String assetTypes();
 
 	@Meta.AD(
 		description = "this-list-is-dynamically-populated",
-		name = "z4h3-asset-types",
-		optionLabels = {"Option 1", "Option 2", "Option 3"},
-		optionValues = {"option1", "option2", "option3"}, required = false
+		name = "dynamically-populated-asset-types",
+		optionLabels = {"Asset1", "Asset2", "Asset3"},
+		optionValues = {"Asset1", "Asset2", "Asset3"}, required = false
 	)
-	String entryList2();
+	String dynamicallyPopulatedAssetTypes();
 
 }
