+++ docs/dxp/latest/en/developing-applications/core-frameworks/configurable-application/providing-dynamic-dropdown-values/resources/liferay-z4h3.zip/z4h3-impl/src/main/java/com/acme/z4h3/internal/configuration/admin/definition/Z4H3ConfigurationFieldOptionsProvider.java
@@ -1,4 +1,4 @@
-package com.acme.z4h3.internal.configuration;
+package com.acme.z4h3.internal.configuration.admin.definition;
 
 import com.liferay.asset.auto.tagger.text.extractor.TextExtractor;
 import com.liferay.asset.auto.tagger.text.extractor.TextExtractorTracker;
@@ -17,7 +17,7 @@ import org.osgi.service.component.annotations.Reference;
 
 @Component(
 	property = {
-		"configuration.field.name=entryList2",
+		"configuration.field.name=dynamicallyPopulatedAssetTypes",
 		"configuration.pid=com.acme.z4h3.internal.configuration.Z4H3Configuration"
 	},
 	service = ConfigurationFieldOptionsProvider.class
