+++ docs/dxp/latest/en/developing-applications/core-frameworks/configurable-application/providing-dynamic-dropdown-values/resources/liferay-z4h3.zip/z4h3-impl/src/main/resources/content/Language.properties
@@ -1,5 +1,8 @@
-z4h3-configuration=Z4H3 Configuration
-z4h3-asset-types=Asset Types
-this-list-is-manually-populated=This list is manually populated.
+asset-types=Asset Types
+dynamically-populated-asset-types=Dynamically Populated Asset Types
 this-list-is-dynamically-populated=This list is dynamically populated.
+this-list-is-manually-populated=This list is manually populated.
+z4h3-configuration-name=Z4H3 Configuration
+
+
